requirements.txt
fastapi==0.73.0
uvicorn==0.15.0
requests
python-dotenv